connect4/game_api/serializers/serializers.py
from rest_framework import serializers
from ..models import Game, Move


class MoveSerializer(serializers.ModelSerializer):
    class Meta:
        model = Move
        fields = ['player', 'column']


class GameSerializer(serializers.ModelSerializer):
    class Meta:
        model = Game
        fields = ['game_id', 'game_over', 'player', 'board']
